[ Warning ] Less compile warnings

+++ src/servers/Scripts/common/Aetheryte.cpp
@@ -19,7 +19,7 @@ public:
    Aetheryte() : EventScript( EVENTSCRIPT_AETHERYTE_ID )
    {}
 
-   void onTalk( uint32_t eventId, Entity::Player& player, uint64_t actorId ) override
+   void onTalk( uint32_t eventId, Entity::Player& player, uint32_t actorId ) override
    {
       if( player.isAetheryteRegistered( eventId & 0xFFFF ) )
       {
